pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lab-data-uploader"
version = "0.1.0"
description = "Reusable local uploader agent for lab PCs — detects completed experiment sessions, generates manifests, deduplicates, and uploads to S3 via presigned URLs."
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.0,<3",
    "pyyaml>=6.0,<7",
    "httpx>=0.27,<1",
    "tenacity>=8.0,<10",
    "structlog>=24.0,<26",
    "typer>=0.12,<1",
]

[project.optional-dependencies]
service = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
    "boto3>=1.34,<2",
]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "respx>=0.21,<1",
    "ruff>=0.5,<1",
    "mypy>=1.10,<2",
    "httpx>=0.27,<1",
]

[project.scripts]
uploader-agent = "agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/service"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
